+++ tests/run_tests.mjs
@@ -85,6 +85,7 @@ const TARGET_ES9 = process.argv.includes('--es9');
 const TARGET_ES10 = process.argv.includes('--es10');
 const TARGET_ES11 = process.argv.includes('--es11');
 const TARGET_ES12 = process.argv.includes('--es12');
+const TARGET_ES13 = process.argv.includes('--es13');
 const TESTS_ONLY = process.argv.includes('-n'); // skip constructing updated test files, dont write anything. Used for code coverage
 const RUN_VERBOSE_IN_SERIAL = process.argv.includes('--serial') || (!SEARCH && !TESTS_ONLY && (INPUT_OVERRIDE || TARGET_FILE || STOP_AFTER_TEST_FAIL || STOP_AFTER_FILE_FAIL));
 const FORCE_WRITE = process.argv.includes('--force-write');
@@ -143,7 +144,7 @@ if (process.argv.includes('-?') || process.argv.includes('--help')) {
     --test-acorn  Also show diff with Acorn AST / pass/fail with test cases (not the same as --acorn !)
     --test-babel  Also show diff with Babel AST / pass/fail with test cases (not the same as --babel !)
     --all         Force to run all four modes (on input)
-    --esX         Where X is one of 6 through 10, like --es6. For -i only, forces the code to run in that version
+    --esX         Where X is one of 6 through 13, like --es6. For -i only, forces the code to run in that version
     --serial      Test all targeted files in serial, verbosely, instead of using parallel phases (which is faster)
                   (Note: -q, -i, and -f implicitly enable --serial)
     --no-printer  Skip running Printer on input
@@ -158,7 +159,7 @@ if (process.argv.includes('-?') || process.argv.includes('--help')) {
   process.exit();
 }
 
-const FORCED_ES_TARGET = TARGET_ES6 ? 6 : TARGET_ES7 ? 7 : TARGET_ES8 ? 8 : TARGET_ES9 ? 9 : TARGET_ES10 ? 10 : TARGET_ES11 ? 11 : TARGET_ES12 ? 12 : undefined;
+const FORCED_ES_TARGET = TARGET_ES6 ? 6 : TARGET_ES7 ? 7 : TARGET_ES8 ? 8 : TARGET_ES9 ? 9 : TARGET_ES10 ? 10 : TARGET_ES11 ? 11 : TARGET_ES12 ? 12 : TARGET_ES13 ? 13 : undefined;
 if (FORCED_ES_TARGET) console.log('Forcing target version: ES' + FORCED_ES_TARGET);
 
 if (AUTO_UPDATE && (AUTO_GENERATE || AUTO_GENERATE_CONSERVATIVE)) throw new Error('Cannot use auto update and auto generate together');